Share common EntityTestHelpers in wemo tests (#62455)
