+++ tests/components/wemo/test_binary_sensor.py
@@ -13,39 +13,7 @@ from homeassistant.components.wemo.binary_sensor import (
 from homeassistant.const import ATTR_ENTITY_ID, STATE_OFF, STATE_ON
 from homeassistant.setup import async_setup_component
 
-from . import entity_test_helpers
-
-
-class EntityTestHelpers:
-    """Common state update helpers."""
-
-    async def test_async_update_locked_multiple_updates(
-        self, hass, pywemo_device, wemo_entity
-    ):
-        """Test that two hass async_update state updates do not proceed at the same time."""
-        await entity_test_helpers.test_async_update_locked_multiple_updates(
-            hass, pywemo_device, wemo_entity
-        )
-
-    async def test_async_update_locked_multiple_callbacks(
-        self, hass, pywemo_device, wemo_entity
-    ):
-        """Test that two device callback state updates do not proceed at the same time."""
-        await entity_test_helpers.test_async_update_locked_multiple_callbacks(
-            hass, pywemo_device, wemo_entity
-        )
-
-    async def test_async_update_locked_callback_and_update(
-        self, hass, pywemo_device, wemo_entity
-    ):
-        """Test that a callback and a state update request can't both happen at the same time.
-
-        When a state update is received via a callback from the device at the same time
-        as hass is calling `async_update`, verify that only one of the updates proceeds.
-        """
-        await entity_test_helpers.test_async_update_locked_callback_and_update(
-            hass, pywemo_device, wemo_entity
-        )
+from .entity_test_helpers import EntityTestHelpers
 
 
 class TestMotion(EntityTestHelpers):
